# server/swagger.json
{
  "openapi": "3.0.2",
  "info": {
...
    "version": "1.0.0",
    "title": "CashStory API (aka Darknight)",
    "termsOfService": "https://cashstory.com/en/legal",
    "contact": {
      "email": "bob@cashstory.com"
    }
  },
...
  "paths": {
...
    "/auth/resetpwd": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset user password",
        "description": "",
        "operationId": "user.resetpwd",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ILogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "500": {
            "description": "no email"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get my user infos",
        "description": "Logged user",
        "operationId": "user.getMe",
        "responses": {
          "200": {
            "description": "Your user data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUser"
                }
              }
            }
          },
          "500": {
            "description": "error"
          }
        }
      },
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Update my user infos",
        "description": "",
        "operationId": "user.updateMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IUser"
              }
            }
          },
          "description": "Updated user object",
          "required": true
        },
        "responses": {
          "400": {
            "description": "error"
          }
        }
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Delete user",
        "description": "Delete my user infos",
        "operationId": "user.deleteMe",
        "responses": {
          "200": {
            "description": "sucess",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/users/me/event": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Event on user",
        "description": "",
        "operationId": "user.addEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "prev": {
                    "type": "object"
                  },
                  "next": {
                    "type": "object"
                  },
                  "meta": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/users/me/favorite": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Add favorite in logged user workspace",
        "description": "",
        "operationId": "user.addFavorite",
        "requestBody": {
          "$ref": "#/components/requestBodies/IFavorite"
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/users/me/favorite/del": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Delete favorite",
        "description": "",
        "operationId": "user.delFavorite",
        "requestBody": {
          "$ref": "#/components/requestBodies/IFavorite"
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all users",
        "description": "",
        "operationId": "user.getAll",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "description": "The numbers of items to return",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IUser"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Insert users",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "user.insert",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IUser"
                }
              }
            }
          },
          "description": "List of user object",
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/users/count": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Count users",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "user.count",
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "user.get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user id that needs to be fetched.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUser"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update user by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "user.update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Updated user id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IUser"
              }
            }
          },
          "description": "Updated user object",
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUser"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "user.delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of user that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/workspaces": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Get all workspace",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.getAll",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "description": "The numbers of items to return",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IWp"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/workspaces/count": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Count workspace",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.count",
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
...
    "/workspaces/{id}": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Get workspace by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IWp"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "put": {
        "tags": [
          "workspace"
        ],
        "summary": "Update workspace by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IWp"
              }
            }
          },
          "description": "Updated workspace object",
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IWp"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "delete": {
        "tags": [
          "workspace"
        ],
        "summary": "Delete workspace by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/section/{sectionId}/box": {
      "post": {
        "tags": [
          "workspace box"
        ],
        "summary": "Add new box in a section",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.addNewBoxToExistingSection",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IBox"
              }
            }
          },
          "description": "Add box object",
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IBox"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/section/{sectionId}/box/{boxId}": {
      "put": {
        "tags": [
          "workspace box"
        ],
        "summary": "Update workspace box by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.updateExistingBox",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "boxId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IBox"
              }
            }
          },
          "description": "Updated box object",
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IBox"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "delete": {
        "tags": [
          "workspace box"
        ],
        "summary": "Delete workspace box by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.deleteBoxFromSection",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "boxId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/smarttables/{workspaceId}/{databaseId}/{collectionId}": {
      "get": {
        "tags": [
          "smarttable"
        ],
        "summary": "Get smarttable",
        "description": "Only works if you are admin",
        "operationId": "smartTable.get",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "databaseId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "collectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ISmartTable"
                }
              }
            }
          },
          "401": {
            "description": "error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BasicAuth": [
        "read",
        "write"
      ]
    },
    {
      "BearerAuth": [
        "read",
        "write"
      ]
    }
  ],
  "servers": [
    {
      "url": "https://{environment}.cashstory.com/api/v1",
      "variables": {
        "environment": {
          "default": "darkknight",
          "enum": [
            "darkknight",
            "darkknight.dev"
          ]
        }
      }
    }
  ],
  "components": {
    "requestBodies": {
      "IFavorite": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/IFavorite"
            }
          }
        },
        "required": true
      }
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ILogin": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "IUser": {
        "type": "object",
        "required": [
          "creatorId",
          "tmpAccount",
          "tmpToken",
          "email",
          "picture",
          "phoneNumber",
          "password",
          "firstName",
          "lastName",
          "emailUrl",
          "searchUrl",
          "company",
          "role",
          "manager",
          "userRole",
          "reset",
          "workspaceCurrent",
          "workspaces",
          "services"
        ],
        "properties": {
          "creatorId": {
            "type": "integer",
            "format": "int64"
          },
          "tmpAccount": {
            "type": "boolean"
          },
          "tmpToken": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "emailUrl": {
            "type": "string"
          },
          "searchUrl": {
            "type": "string"
          },
          "company": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/ICompany"
              }
            }
          },
          "role": {
            "type": "string"
          },
          "manager": {
            "type": "string"
          },
          "userRole": {
            "type": "string"
          },
          "reset": {
            "type": "boolean"
          },
          "workspaceCurrent": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IWp"
              }
            }
          },
          "workspaces": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IWorkspaceList"
              }
            }
          },
          "services": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IServiceList"
              }
            }
          }
        },
        "xml": {
          "name": "User"
        }
      },
      "IWp": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "section": {
            "type": "integer",
            "format": "int32"
          },
          "box": {
            "type": "string"
          },
          "logo": {
            "type": "object",
            "items": {
              "$ref": "#/components/schemas/ILogo"
            }
          }
        },
        "xml": {
          "name": "IWp"
        }
      },
      "IBox": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "urlBg"
        ],
        "properties": {
          "_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "urlBg": {
            "type": "string"
          },
          "iframe": {
            "type": "boolean"
          },
          "hideElements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "login": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/ILogin"
              }
            }
          },
          "smartTable": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/ISmartTable"
              }
            }
          },
          "cors": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "buttons": {
            "type": "object",
            "properties": {
              "notebookUrl": {
                "type": "string"
              },
              "expandCollapse": {
                "type": "boolean"
              },
              "openNewWindow": {
                "type": "boolean"
              },
              "backTo": {
                "type": "boolean"
              }
            }
          },
          "authMethod": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "autoExpand": {
            "type": "boolean"
          }
        },
        "xml": {
          "name": "IBox"
        }
      },
...
      "ISmartTable": {
        "type": "object",
        "required": [
          "database",
          "collection",
          "settings"
        ],
        "properties": {
          "database": {
            "type": "string"
          },
          "collection": {
            "type": "string"
          },
          "settings": {
            "type": "object"
          },
          "model": {
            "type": "string"
          },
          "rights": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IRights"
              }
            }
          }
        },
        "xml": {
          "name": "ISmartTable"
        }
      },
      "ICompany": {
        "type": "object",
        "required": [
          "name",
          "size",
          "whatUse",
          "kind"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "whatUse": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          }
        },
        "xml": {
          "name": "ICompany"
        }
      },
      "IWorkspaceList": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IWorkspaceConfig"
              }
            }
          }
        },
        "xml": {
          "name": "IWorkspaceList"
        }
      },
...
      "IServiceList": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IService"
              }
            }
          }
        }
      },
...
      "ILogo": {
        "type": "object",
        "required": [
          "url",
          "name"
        ],
        "properties": {
          "url": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}